# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend monolith modular untuk platform jasa informal Kerjantara.id. API ini menghubungkan employer dengan worker melalui matching engine berbasis GPS dan skor reputasi.",
        "title": "Kerjantara Backend API",
        "contact": {},
        "version": "v3.3-hackathon"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/payments/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Employer membuat pembayaran escrow melalui Midtrans Snap. Logika platform fee: transaksi di bawah Rp 1.000.000 dikenakan flat Rp 10.000; transaksi Rp 1.000.000 ke atas dikenakan 2% dari agreed_price.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Buat pembayaran untuk job yang sudah diterima",
                "parameters": [
                    {
                        "description": "Job ID yang akan dibayar",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_payment.CreatePaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docs.SuccessEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_payment.CreatePaymentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorEnvelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/payments/webhook": {
            "post": {
                "description": "Endpoint ini dipanggil otomatis oleh server Midtrans setelah transaksi berhasil atau gagal. Tidak memerlukan autentikasi JWT.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Webhook notifikasi pembayaran dari Midtrans",
                "parameters": [
                    {
                        "description": "Payload notifikasi Midtrans",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_payment.WebhookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/docs.SuccessEnvelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorEnvelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/payments/{job_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengembalikan detail pembayaran escrow untuk job tertentu, termasuk status escrow (pending, held, released, refunded), amount, platform_fee, dan net_to_worker.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Cek status pembayaran berdasarkan Job ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID (UUID)",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/docs.SuccessEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/payments/{job_id}/milestones": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengembalikan array milestone pembayaran harian untuk job multi-hari (duration_days \u003e 1). Setiap milestone merepresentasikan pembayaran per hari yang dirilis secara bertahap setelah employer mengkonfirmasi penyelesaian hari tersebut.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Ambil daftar milestone pembayaran harian",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID (UUID)",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/docs.SuccessEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
        "/workers/{id}/score": {
            "get": {
                "description": "Mengembalikan KerjantaraScore (rata-rata rating 0.00–5.00) beserta riwayat perubahan skor. kerjantara_score dihitung sebagai rata-rata dari semua rating yang diterima pekerja dari tabel trx_ratings, diperbarui otomatis saat event job.rated diterima. Endpoint ini bersifat publik dan tidak memerlukan autentikasi.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Score"
                ],
                "summary": "Lihat skor reputasi pekerja",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Worker ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docs.SuccessEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_score.WorkerScoreData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorEnvelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "docs.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "message": {
                    "type": "string",
                    "example": "Field tidak valid"
                }
            }
        },
        "docs.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/docs.ErrorDetail"
                }
            }
        },
        "docs.MetaInfo": {
            "type": "object",
            "properties": {
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-15T08:30:00+07:00"
                }
            }
        },
        "docs.SuccessEnvelope": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/docs.MetaInfo"
                }
            }
        },
...
        "internal_payment.CreatePaymentRequest": {
            "type": "object",
            "properties": {
                "job_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "internal_payment.CreatePaymentResponse": {
            "type": "object",
            "properties": {
                "agreed_price": {
                    "type": "integer",
                    "example": 500000
                },
                "fee_note": {
                    "type": "string",
                    "example": "Biaya layanan Rp 10.000 (transaksi di bawah Rp 1.000.000)"
                },
                "midtrans_order_id": {
                    "type": "string",
                    "example": "order-550e8400-1234"
                },
                "net_to_worker": {
                    "type": "integer",
                    "example": 490000
                },
                "payment_id": {
                    "type": "string",
                    "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                },
                "platform_fee": {
                    "type": "integer",
                    "example": 10000
                },
                "snap_token": {
                    "type": "string",
                    "example": "snap-token-abc123"
                },
                "total_charged_to_employer": {
                    "type": "integer",
                    "example": 500000
                }
            }
        },
        "internal_payment.WebhookRequest": {
            "type": "object",
            "properties": {
                "fraud_status": {
                    "type": "string",
                    "example": "accept"
                },
                "order_id": {
                    "type": "string",
                    "example": "order-550e8400-1234"
                },
                "transaction_status": {
                    "type": "string",
                    "example": "settlement"
                }
            }
        },
        "internal_score.ScoreHistoryItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "delta": {
                    "type": "number"
                },
                "job_id": {
                    "type": "string"
                },
                "score_after": {
                    "type": "number"
                },
                "score_before": {
                    "type": "number"
                }
            }
        },
        "internal_score.WorkerScoreData": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string"
                },
                "history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_score.ScoreHistoryItem"
                    }
                },
                "kerjantara_score": {
                    "type": "number"
                },
                "total_jobs_done": {
                    "type": "integer"
                },
                "worker_id": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token JWT dalam format: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}